More subtle color change on drag, reduce vertical height
